Scripts/requirements.txt
numpy==1.22.4
pandas==1.5.3
pandas-datareader==0.10.0
pandas-gbq==0.17.9
Pillow==8.4.0
tensorflow==2.12.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.12.0
tensorflow-gcs-config==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-probability==0.19.0
